[25490] Add publicity settings for query saving modal
When copying queries, the publicity for other members is copied, but not
the query `star` setting (whether it is shown in the project menu or not).

This is confusing to the user in two ways:

1. The user does not know whether a query is public or not when copying
it

2. It is weird to copy one setting, but not the other

This PR suggests to extend the `save as` modal with the visibility
settings. To this end, I extracted the visibility form into a separate
directive to be re-used from the publish/sharing modal and the save as
modal.

As://community.openproject.com/projects/openproject/work_packages/25490